requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract==0.3.10
Pillow>=10.0.0
easyocr>=1.7.0 